LayoutTests/imported/w3c/web-platform-tests/css/css-images/parsing/gradient-position-invalid-expected.txt


PASS e.style['background-image'] = "radial-gradient(at top 0px, red, blue)" should not set the property value 
FAIL e.style['background-image'] = "radial-gradient(at center left 1px, red, blue)" should not set the property value assert_equals: expected "" but got "radial-gradient(at left 1px center, red, blue)"
FAIL e.style['background-image'] = "radial-gradient(at center top 2px, red, blue)" should not set the property value assert_equals: expected "" but got "radial-gradient(at center top 2px, red, blue)"
FAIL e.style['background-image'] = "radial-gradient(at right 3% center, red, blue)" should not set the property value assert_equals: expected "" but got "radial-gradient(at right 3% center, red, blue)"
FAIL e.style['background-image'] = "radial-gradient(at left 4px top, red, blue)" should not set the property value assert_equals: expected "" but got "radial-gradient(at left 4px top, red, blue)"
FAIL e.style['background-image'] = "radial-gradient(at right top 5px, red, blue)" should not set the property value assert_equals: expected "" but got "radial-gradient(at right top 5px, red, blue)"
FAIL e.style['background-image'] = "radial-gradient(at bottom 6% center, red, blue)" should not set the property value assert_equals: expected "" but got "radial-gradient(at center bottom 6%, red, blue)"
FAIL e.style['background-image'] = "radial-gradient(at bottom 7% left, red, blue)" should not set the property value assert_equals: expected "" but got "radial-gradient(at left bottom 7%, red, blue)"
FAIL e.style['background-image'] = "radial-gradient(at bottom right 8%, red, blue)" should not set the property value assert_equals: expected "" but got "radial-gradient(at right 8% bottom, red, blue)"
